Refactor features, enable build with --no-default-features
