+++ .github/workflows/main.yml
@@ -191,11 +191,11 @@ jobs:
       - name: "Verify aries_vcx compiles with different dependency feature variants"
         run: |
           cd aries_vcx
-          cargo check
+          cargo check --no-default-features
           cargo check --tests --all-features
-          cargo check --features vdrtools
-          cargo check --features modular_libs
-          cargo check --features vdr_proxy_ledger
+          cargo check --features vdrtools --no-default-features
+          cargo check --features modular_libs --no-default-features
+          cargo check --features vdr_proxy_ledger --no-default-features
 
   ##########################################################################################
   ##############################   DOCKER BUILD   ##########################################
