+++ aries_vcx/Cargo.toml
@@ -12,7 +12,7 @@ doctest = false
 
 [features]
 default = ["vdrtools"]
-vdrtools = ["aries_vcx_core/vdrtools"]
+vdrtools = ["aries_vcx_core/vdrtools_anoncreds", "aries_vcx_core/vdrtools_ledger", "aries_vcx_core/vdrtools_wallet"]
 # Feature flag to include the 'modular library' dependencies (vdrtools alternatives; indy-vdr, indy-credx)
 modular_libs = ["aries_vcx_core/modular_libs"]
 # TODO: Remove using "vdrtools" feature flag for vdr_proxy_ledger once IndyCredxAnonCreds
