+++ aries_vcx/tests/utils/migration.rs
@@ -8,7 +8,7 @@ use aries_vcx::{
     core::profile::modular_libs_profile::ModularLibsProfile, global::settings::WALLET_KDF_RAW,
     utils::devsetup::SetupProfile,
 };
-use aries_vcx_core::wallet::indy::agency_client_wallet::ToBaseAgencyClientWallet;
+use aries_vcx_core::wallet::agency_client_wallet::ToBaseAgencyClientWallet;
 use aries_vcx_core::wallet::indy::wallet::create_and_open_wallet;
 use aries_vcx_core::wallet::indy::IndySdkWallet;
 use aries_vcx_core::wallet::indy::WalletConfig;
