+++ aries_vcx_core/src/anoncreds/mod.rs
@@ -3,7 +3,7 @@ pub mod base_anoncreds;
 #[cfg(any(feature = "modular_libs", feature = "vdr_proxy_ledger"))]
 pub mod credx_anoncreds;
 
-#[cfg(feature = "vdrtools")]
+#[cfg(feature = "vdrtools_anoncreds")]
 pub mod indy;
-#[cfg(feature = "vdrtools")]
+#[cfg(feature = "vdrtools_anoncreds")]
 pub mod indy_anoncreds;
