+++ aries_vcx_core/src/errors/mapping_vdrtools.rs
@@ -1,17 +1,16 @@
-#[cfg(feature = "vdrtools")]
 use vdrtools::types::{
     errors::{IndyError, IndyErrorKind},
     ErrorCode,
 };
 
-#[cfg(all(
-    not(feature = "vdrtools"),
-    any(feature = "modular_libs", feature = "vdr_proxy_ledger")
-))]
-use indy_ledger_response_parser::{
-    errors::{IndyError, IndyErrorKind},
-    ErrorCode,
-};
+// #[cfg(all(
+//     not(feature = "vdrtools"),
+//     any(feature = "modular_libs", feature = "vdr_proxy_ledger")
+// ))]
+// use indy_ledger_response_parser::{
+//     errors::{IndyError, IndyErrorKind},
+//     ErrorCode,
+// };
 
 use crate::errors::error::{AriesVcxCoreError, AriesVcxCoreErrorKind};
