+++ aries_vcx_core/src/errors/mod.rs
@@ -7,5 +7,5 @@ mod mapping_indyvdr;
 #[cfg(feature = "vdr_proxy_ledger")]
 mod mapping_indyvdr_proxy;
 mod mapping_others;
-#[cfg(any(feature = "vdrtools", feature = "modular_libs", feature = "vdr_proxy_ledger"))]
+#[cfg(any(feature = "vdrtools_wallet", feature = "vdrtools_ledger", feature = "vdrtools_anoncreds", feature = "vdr_proxy_ledger"))]
 mod mapping_vdrtools;
