+++ aries_vcx_core/src/wallet/base_wallet.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use async_trait::async_trait;
 
-#[cfg(feature = "vdrtools")]
+#[cfg(feature = "vdrtools_wallet")]
 use vdrtools::WalletHandle;
 
 use crate::errors::error::VcxCoreResult;
@@ -13,7 +13,7 @@ use crate::utils::async_fn_iterator::AsyncFnIterator;
 /// see: <https://github.com/hyperledger/indy-sdk/blob/main/libindy/src/api/wallet.rs>
 #[async_trait]
 pub trait BaseWallet: std::fmt::Debug + Send + Sync {
-    #[cfg(feature = "vdrtools")]
+    #[cfg(feature = "vdrtools_wallet")]
     fn get_wallet_handle(&self) -> WalletHandle;
 
     // ----- DIDs
