+++ aries_vcx_core/src/wallet/indy/mod.rs
@@ -15,7 +15,6 @@ use crate::{SearchHandle, WalletHandle};
 
 use super::base_wallet::BaseWallet;
 
-pub mod agency_client_wallet;
 pub mod did_mocks;
 pub mod indy_wallet;
 pub mod internal;
