+++ aries_vcx_core/src/wallet/mod.rs
@@ -1,3 +1,4 @@
+pub mod agency_client_wallet;
 pub mod base_wallet;
-#[cfg(feature = "vdrtools")]
+#[cfg(feature = "vdrtools_wallet")]
 pub mod indy;
